+++ packages/integration-test/importDefault/observableFromAsync.test.ts
@@ -1,0 +1,20 @@
+import { observableFromAsync } from 'iter-fest';
+
+test('observableFromAsync should work', async () => {
+  const observable = observableFromAsync(
+    (async function* () {
+      yield 1;
+      yield 2;
+      yield 3;
+    })()
+  );
+
+  const next = jest.fn();
+
+  await new Promise<void>(resolve => observable.subscribe({ complete: resolve, next }));
+
+  expect(next).toHaveBeenCalledTimes(3);
+  expect(next).toHaveBeenNthCalledWith(1, 1);
+  expect(next).toHaveBeenNthCalledWith(2, 2);
+  expect(next).toHaveBeenNthCalledWith(3, 3);
+});
