+++ packages/integration-test/importDefault/symbolObservable.test.ts
@@ -1,0 +1,7 @@
+import { Observable, SymbolObservable } from 'iter-fest';
+
+test('SymbolObservable should work', () => {
+  const observable = new Observable(() => {});
+
+  expect(observable[SymbolObservable]()).toBe(observable);
+});
