+++ packages/integration-test/importNamed/observable.test.ts
@@ -1,0 +1,17 @@
+import { Observable } from 'iter-fest/observable';
+
+test('Observable should work', () => {
+  const next = jest.fn();
+  const complete = jest.fn();
+
+  const observable = new Observable(observer => {
+    observer.next(1);
+    observer.complete();
+  });
+
+  observable.subscribe({ complete, next });
+
+  expect(next).toHaveBeenCalledTimes(1);
+  expect(next).toHaveBeenNthCalledWith(1, 1);
+  expect(complete).toHaveBeenCalledTimes(1);
+});
