+++ packages/integration-test/importNamed/observableSubscribeAsReadable.test.ts
@@ -1,0 +1,22 @@
+import { Observable } from 'iter-fest/observable';
+import { observableSubscribeAsReadable } from 'iter-fest/observableSubscribeAsReadable';
+
+test('observableSubscribeAsReadable should work', async () => {
+  const stream = new TextDecoderStream();
+  const observable = Observable.from([65, 66, 67]);
+  const readable = observableSubscribeAsReadable(observable);
+  const numberToInt8Array = new TransformStream({
+    transform(chunk, controller) {
+      controller.enqueue(Int8Array.from([chunk]));
+    }
+  });
+
+  readable.pipeThrough(numberToInt8Array).pipeTo(stream.writable);
+
+  const reader = stream.readable.getReader();
+
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 'A' });
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 'B' });
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 'C' });
+  await expect(reader.read()).resolves.toEqual({ done: true });
+});
