+++ packages/integration-test/requireDefault/iteratorToIterable.test.ts
@@ -1,0 +1,31 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { iteratorToIterable, Observable } = require('iter-fest');
+
+test('iteratorToIterable should work', () =>
+  expect(
+    Array.from(
+      iteratorToIterable(
+        ((): Iterator<number> => {
+          let value = 0;
+
+          return { next: () => (++value <= 3 ? { done: false, value } : { done: true, value: undefined }) };
+        })()
+      )
+    )
+  ).toEqual([1, 2, 3]));
+
+test('Observable should work', () => {
+  const next = jest.fn();
+  const complete = jest.fn();
+
+  const observable = new Observable(observer => {
+    observer.next(1);
+    observer.complete();
+  });
+
+  observable.subscribe({ complete, next });
+
+  expect(next).toHaveBeenCalledTimes(1);
+  expect(next).toHaveBeenNthCalledWith(1, 1);
+  expect(complete).toHaveBeenCalledTimes(1);
+});
