+++ packages/integration-test/requireDefault/readableStreamFrom.test.ts
@@ -1,0 +1,13 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { readableStreamFrom } = require('iter-fest');
+
+test('readableStreamFrom should work', async () => {
+  const iterable = [1, 2, 3].values();
+
+  const reader = readableStreamFrom(iterable).getReader();
+
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 1 });
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 2 });
+  await expect(reader.read()).resolves.toEqual({ done: false, value: 3 });
+  await expect(reader.read()).resolves.toEqual({ done: true, value: undefined });
+});
