+++ packages/integration-test/requireDefault/symbolObservable.test.ts
@@ -1,0 +1,8 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { Observable, SymbolObservable } = require('iter-fest');
+
+test('SymbolObservable should work', () => {
+  const observable = new Observable(() => {});
+
+  expect(observable[SymbolObservable]()).toBe(observable);
+});
