+++ packages/integration-test/requireNamed/symbolObservable.test.ts
@@ -1,0 +1,9 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { Observable } = require('iter-fest/observable');
+const { SymbolObservable } = require('iter-fest/symbolObservable');
+
+test('SymbolObservable should work', () => {
+  const observable = new Observable(() => {});
+
+  expect(observable[SymbolObservable]()).toBe(observable);
+});
